# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Issue Tracker API
  description: REST API for a Redmine-like issue tracker. Supports projects, issues, comments, and role-based access control.
  license:
    name: ''
  version: 0.1.0
paths:
...
  /api/projects:
    get:
      tags:
      - Projects
      summary: List projects
      description: Returns all projects visible to the authenticated user (public + owned/member).
      operationId: list_projects
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Paginated list of projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProjectResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer: []
    post:
      tags:
      - Projects
      summary: Create project
      description: Creates a new project. The authenticated user becomes the owner.
      operationId: create_project
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
        required: true
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Identifier already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer: []
  /api/projects/{project_id}:
    get:
      tags:
      - Projects
      summary: Get project
      description: Returns a single project by ID. Requires visibility access.
      operationId: get_project
      parameters:
      - name: project_id
        in: path
        description: Project ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer: []
    delete:
      tags:
      - Projects
      summary: Delete project
      description: Deletes a project. Only the project owner can perform this action.
      operationId: delete_project
      parameters:
      - name: project_id
        in: path
        description: Project ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Project deleted
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied (not owner)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer: []
    patch:
      tags:
      - Projects
      summary: Update project
      description: Updates a project's name, description, or visibility. Requires owner or admin role.
      operationId: update_project
      parameters:
      - name: project_id
        in: path
        description: Project ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
        required: true
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer: []
...
  /api/projects/{project_id}/members:
    get:
      tags:
      - Projects
      summary: List project members
      description: Returns the members of a project. Requires visibility access to the project.
      operationId: list_members
      parameters:
      - name: project_id
        in: path
        description: Project ID
        required: true
        schema:
          type: string
          format: uuid
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Paginated list of members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProjectMemberResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer: []
    post:
      tags:
      - Projects
      summary: Add project member
      description: Adds a user to a project with the specified role. Requires owner or admin role.
      operationId: add_member
      parameters:
      - name: project_id
        in: path
        description: Project ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemberRequest'
        required: true
      responses:
        '204':
          description: Member added
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User is already a member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer: []
  /api/users/me:
    get:
      tags:
      - Users
      summary: Get current user
      description: Returns the authenticated user's profile.
      operationId: get_me
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer: []
    patch:
      tags:
      - Users
      summary: Update current user
      description: Updates the authenticated user's display name and/or password. Only supplied fields are changed.
      operationId: update_me
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
        required: true
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer: []
components:
  schemas:
    AddMemberRequest:
      type: object
      description: Request body for `POST /api/projects/:project_id/members`.
      required:
      - user_id
      - role
      properties:
        role:
          $ref: '#/components/schemas/MemberRole'
        user_id:
          type: string
          format: uuid
...
    CreateProjectRequest:
      type: object
      description: Request body for `POST /api/projects`.
      required:
      - name
      - identifier
      properties:
        description:
          type: string
          default: ''
        identifier:
          type: string
          description: |-
            URL-safe slug that uniquely identifies the project (e.g., "my-project").
            Allowed characters: lowercase letters, digits, and hyphens.
            Must start and end with a letter or digit.
        is_public:
          type: boolean
          default: true
        name:
          type: string
    ErrorResponse:
      type: object
      description: |-
        Schema-only type for documenting the standard error response shape in OpenAPI.
        Not used at runtime — the actual JSON is built inline in `AppError::into_response`.
      required:
      - message
      properties:
        data:
          description: Additional error details (validation field errors, etc.) or null.
        message:
          type: string
          description: Human-readable error message.
...
    MemberRole:
      type: string
      enum:
      - admin
      - member
      - viewer
...
    PaginatedProjectMemberResponse:
      type: object
      required:
      - data
      - total
      - page
      - limit
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProjectMember'
        limit:
          type: integer
          format: int64
        page:
          type: integer
          format: int64
        total:
          type: integer
          format: int64
    PaginatedProjectResponse:
      type: object
      required:
      - data
      - total
      - page
      - limit
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        limit:
          type: integer
          format: int64
        page:
          type: integer
          format: int64
        total:
          type: integer
          format: int64
    Project:
      type: object
      required:
      - id
      - name
      - identifier
      - description
      - is_public
      - owner_id
      - created_at
      - updated_at
      properties:
        created_at:
          type: string
          format: date-time
        description:
          type: string
        id:
          type: string
          format: uuid
        identifier:
          type: string
        is_public:
          type: boolean
        name:
          type: string
        owner_id:
          type: string
          format: uuid
        updated_at:
          type: string
          format: date-time
    ProjectMember:
      type: object
      required:
      - project_id
      - user_id
      - role
      - joined_at
      properties:
        joined_at:
          type: string
          format: date-time
        project_id:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/MemberRole'
        user_id:
          type: string
          format: uuid
...
    UpdateProjectRequest:
      type: object
      description: |-
        Request body for `PATCH /api/projects/:project_id`.

        All fields optional. Omitted fields are left unchanged.
      properties:
        description:
          type:
          - string
          - 'null'
        is_public:
          type:
          - boolean
          - 'null'
        name:
          type:
          - string
          - 'null'
    UpdateUserRequest:
      type: object
      description: |-
        Request body for `PATCH /api/users/me`.

        All fields are optional — only provided fields are updated.
      properties:
        display_name:
          type:
          - string
          - 'null'
        new_password:
          type:
          - string
          - 'null'
    UserResponse:
      type: object
      description: Safe view of a user — everything except the password hash.
      required:
      - id
      - email
      - display_name
      - created_at
      - updated_at
      properties:
        created_at:
          type: string
          format: date-time
        display_name:
          type: string
        email:
          type: string
        id:
          type: string
          format: uuid
        updated_at:
          type: string
          format: date-time
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...